+++ src/main/java/demo/example/pageandsort/Product.java
@@ -4,9 +4,9 @@ public class Product {
 
     private String name;
 
-    private long price;
+    private int price;
 
-    public Product(String name, long price) {
+    public Product(String name, int price) {
         this.name = name;
         this.price = price;
     }
